added SLC, Go to tims squares, Makefile and changed board and residence a bit

+++ needlesrng.h
@@ -5,7 +5,7 @@
 
 class NeedlesRng : public Factory {
     std::vector<float> rngRange; // breakpoints that mark right endpoint of each event 
-    std::uniform_real_distribution<int> distribution; // range of random values that can be generated (0.0 - 1.0)
+    std::uniform_real_distribution<float> distribution; // range of random values that can be generated (0.0 - 1.0)
     public:
         NeedlesRng();
         virtual Event generateEvent() override;
